+++ src/main/java/service/DepositOperations.java
@@ -2,6 +2,8 @@ package service;
 
 import model.BankAccount;
 
+import java.math.BigDecimal;
+
 public interface DepositOperations {
-    void deposit(BankAccount bankAccount, double amount);
+    void deposit(BankAccount bankAccount, BigDecimal amount);
 }
