+++ src/main/java/service/StatementService.java
@@ -4,12 +4,13 @@ import model.AccountStatement;
 import model.BankAccount;
 import model.BankOperationType;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 public class StatementService implements BankStatementService {
 
     @Override
-    public void updateAccountStatement(BankAccount bankAccount, double amount, BankOperationType operation, ZonedDateTime zonedDateTime) {
+    public void updateAccountStatement(BankAccount bankAccount, BigDecimal amount, BankOperationType operation, ZonedDateTime zonedDateTime) {
         var accountStatement = new AccountStatement(operation, zonedDateTime, amount, bankAccount.getBalance());
         bankAccount.getAccountStatements().add(accountStatement);
     }
