+++ src/main/java/service/WithdrawalOperations.java
@@ -2,7 +2,9 @@ package service;
 
 import model.BankAccount;
 
+import java.math.BigDecimal;
+
 public interface WithdrawalOperations {
 
-    void withdrawal(BankAccount bankAccount, double amount);
+    void withdrawal(BankAccount bankAccount, BigDecimal amount);
 }
